+++ cmd/demo.go
@@ -51,6 +51,16 @@ var cmdDemoPwd = &cobra.Command{
 	},
 }
 
+var cmdDemoDelete = &cobra.Command{
+	Use:   "delete",
+	Short: "Delete the demo Kubernetes cluster.",
+	Long: `Delete the demo Kubernetes cluster in order to free corresponding 
+	resources.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		demo.DeleteKubernetesCluster()
+	},
+}
+
 func ExecuteA8sPGDemo() {
 	demo.PrintWelcomeScreen()
 
@@ -78,12 +88,18 @@ func ExecuteA8sPGDemo() {
 }
 
 func init() {
-	cmdDemoA8sPG.PersistentFlags().StringVarP(&demo.KubernetesTool, "provider", "p", "minikube", "provider for creating the Kubernetes cluster. Valid options are \"minikube\" an \"kind\"")
+
+	// Depricated: Focus on minikube support
+	//cmdDemoA8sPG.PersistentFlags().StringVarP(&demo.KubernetesTool, "provider", "p", "minikube", "provider for creating the Kubernetes cluster. Valid options are \"minikube\" an \"kind\"")
+
 	cmdDemoA8sPG.PersistentFlags().StringVar(&demo.BackupInfrastructureRegion, "backup-region", "us-east-1", "specify the infrastructure region to store backups such as \"us-east-1\".")
 	cmdDemoA8sPG.PersistentFlags().StringVar(&demo.BackupInfrastructureBucket, "backup-bucket", "a8s-backups", "specify the infrastructure object store bucket name.")
 	cmdDemoA8sPG.PersistentFlags().StringVar(&demo.BackupInfrastructureBucket, "backup-provider", "AWS", "specify the infrastructure provider as supported by the a8s Backup Manager.")
 	cmdDemo.AddCommand(cmdDemoA8sPG)
+	cmdDemo.PersistentFlags().StringVarP(&demo.DemoClusterName, "cluster-name", "c", "a8s-demo", "name of the demo Kubernetes cluster.")
+	cmdDemo.PersistentFlags().BoolVarP(&demo.UnattendedMode, "yes", "y", false, "skip yes-no questions by answering with \"yes\".")
 	cmdDemo.AddCommand(cmdDemoPwd)
+	cmdDemo.AddCommand(cmdDemoDelete)
 	rootCmd.AddCommand(cmdDemo)
 
 	// Here you will define your flags and configuration settings.